review and question finished

+++ src/pages/review/review_answer.js
@@ -6,7 +6,7 @@ import Button from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
 import Form from "react-bootstrap/Form";
 import "../../styles/review.css";
-import { useNavigate, useParams } from "react-router-dom";
+import { Link, useNavigate, useParams } from "react-router-dom";
 import {
   addDoc,
   arrayRemove,
@@ -16,7 +16,9 @@ import {
   doc,
   getDoc,
   getDocs,
+  limit,
   onSnapshot,
+  orderBy,
   query,
   updateDoc,
   where,
@@ -68,12 +70,6 @@ function Qmodal() {
     }
     checkInfo();
   }, [selectedOption]);
-
-  function checkLogin() {
-    if (currentUser === null) {
-      navigate("/");
-    }
-  }
 
   async function createData(postData) {
     try {
@@ -558,6 +554,7 @@ const Answer = ({ userData }) => {
                           postID={id}
                           rep_users={post.rep_users}
                           rep_count={post.report}
+                          tagName={post.tag}
                           member_id={post.member_id}
                         />
                       </Dropdown>
@@ -617,7 +614,7 @@ const Answer = ({ userData }) => {
                   {comment ? (
                     <div>
                       {comment.map((item) => (
-                        <div key={item.id}>
+                        <div className="post-border" key={item.id}>
                           <div className="flex-container comment pt-3">
                             <MemberInfo memberID={item.member_id} />
 
@@ -689,9 +686,10 @@ const Answer = ({ userData }) => {
                               />
                             </div>
                           </div>
-                          <hr />
+
                           {visibility[item.id] && (
                             <div className="container">
+                              <hr />
                               <ReplyLoad
                                 userData={userData}
                                 postID={id}
@@ -766,14 +764,7 @@ const Answer = ({ userData }) => {
                   <span className="">เป็นที่นิยมใน Mod's Talk</span>
                 </div>
 
-                <Button className="hit-tag">Programming</Button>
-                <Button className="hit-tag">Data Science</Button>
-                <Button className="hit-tag">Technology</Button>
-                <Button className="hit-tag">Self Improvement</Button>
-                <Button className="hit-tag">Writing</Button>
-                <Button className="hit-tag">Ralationships</Button>
-                <Button className="hit-tag">Machine Learning</Button>
-                <Button className="hit-tag">Some random topic</Button>
+                <PopularTag />
               </div>
             </div>
           </div>
@@ -909,6 +900,43 @@ function MemberInfo({ memberID }) {
           <div style={{ paddingLeft: "1rem" }}></div>
         </div>
       )}
+    </div>
+  );
+}
+
+function PopularTag({}) {
+  const [data, setData] = useState([]); // initialize state variable for data
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const docRef = collection(db, "tag_ranked"); // create a reference to the "tag_ranked" collection
+
+        const q = query(docRef, orderBy("count", "desc"), limit(10)); // create a query that sorts by "count" field in ascending order and limits to 5 documents
+
+        const snapshot = await getDocs(q); // execute the query and get the snapshot of results
+
+        const documentsData = snapshot.docs.map((doc) => ({
+          id: doc.id,
+          ...doc.data(),
+        })); // map through each document and extract its data and ID
+
+        setData(documentsData); // set the state variable to the retrieved data
+      } catch (error) {
+        console.error("Error fetching document: ", error);
+      }
+    };
+
+    fetchData();
+  }, []);
+
+  return (
+    <div className="pb-5">
+      {data.map((doc) => (
+        <Link to={`/review/tag/${doc.id}`}>
+          <Button className="hit-tag">{doc.id}</Button>
+        </Link>
+      ))}
     </div>
   );
 }
@@ -1114,16 +1142,13 @@ function ReplyLoad({ userData, postID, cmntID, replyCount }) {
         </div>
       </form>
       {/* เขียนคอมเมนท์                           */}
-      <hr />
     </div>
   );
 }
 
-function Rep_Del_Click({ postID, rep_users, rep_count, member_id }) {
+function Rep_Del_Click({ postID, rep_users, rep_count, tagName, member_id }) {
   const currentUser = auth.currentUser;
   const currentUserId = currentUser.uid;
-
-  const navigate = useNavigate();
 
   const [authorCheck, setAuthorCheck] = useState(false);
   const [reportedByCurrentUser, setReportedByCurrentUser] = useState(false);
@@ -1135,89 +1160,121 @@ function Rep_Del_Click({ postID, rep_users, rep_count, member_id }) {
       } else {
         setReportedByCurrentUser(false);
       }
-
-      if (member_id === currentUserId) {
-        setAuthorCheck(true);
-      } else {
-        setAuthorCheck(false);
-      }
     } catch (e) {
       console.error(e);
     }
   }, [rep_users, reportedByCurrentUser]);
 
+  useEffect(() => {
+    if (member_id === currentUserId) {
+      setAuthorCheck(false);
+    } else if (member_id !== currentUserId) {
+      setAuthorCheck(true);
+    }
+  }, []);
+
   const handleReportClick = () => {
-    const docRef = doc(db, "review", postID);
-    if (reportedByCurrentUser === false) {
-      updateDoc(docRef, {
-        rep_users: arrayUnion(currentUserId),
-        report: rep_count + 1,
-      })
-        .then(() => {
-          console.log("You Report the post!");
+    const confirmed = window.confirm(
+      "คุณยืนยันที่ต้องการจะรายงานโพสต์ใช่หรือไม่ ?"
+    );
+
+    if (confirmed) {
+      const docRef = doc(db, "review", postID);
+      if (reportedByCurrentUser === false) {
+        updateDoc(docRef, {
+          rep_users: arrayUnion(currentUserId),
+          report: rep_count + 1,
         })
-        .catch((error) => {
-          console.error("Error updating document: ", error);
-        });
-      alert("ขอบคุณที่แจ้งรายงาน ทางแอดมินจะพยายามตรวจสอบให้เร็วที่สุด");
-    } else if (reportedByCurrentUser === true) {
-      alert("คุณได้รายงานโพสต์ไปแล้ว");
+          .then(() => {
+            console.log("You Report the post!");
+          })
+          .catch((error) => {
+            console.error("Error updating document: ", error);
+          });
+        alert("ขอบคุณที่แจ้งรายงาน ทางแอดมินจะพยายามตรวจสอบให้เร็วที่สุด");
+      } else if (reportedByCurrentUser === true) {
+        alert("คุณได้รายงานโพสต์ไปแล้ว");
+      }
     }
   };
 
   const handleDeleteClick = async () => {
-    // Create a reference to the post document
-    const docRef = doc(db, "review", postID);
-
-    // Create a query to get all comments for the post
-    const commentQuery = query(
-      collection(db, "cmnt_review"),
-      where("post_id", "==", postID)
+    const confirmed = window.confirm(
+      "คุณยืนยันที่ต้องการจะลบโพสต์ใช่หรือไม่ ?"
     );
 
-    // Create a query to get all replies for the post
-    const replyQuery = query(
-      collection(db, "reply_review"),
-      where("post_id", "==", postID)
-    );
+    if (confirmed) {
+      // Create a reference to the post document
+      const docRef = doc(db, "review", postID);
 
-    // Use Promise.all() to execute both queries in parallel
-    const [commentSnapshot, replySnapshot] = await Promise.all([
-      getDocs(commentQuery),
-      getDocs(replyQuery),
-    ]);
+      // Create a query to get all comments for the post
+      const commentQuery = query(
+        collection(db, "cmnt_review"),
+        where("post_id", "==", postID)
+      );
 
-    // Delete the post document
-    await deleteDoc(docRef);
+      // Create a query to get all replies for the post
+      const replyQuery = query(
+        collection(db, "reply_review"),
+        where("post_id", "==", postID)
+      );
 
-    // Delete all comment documents
-    commentSnapshot.forEach(async (commentDoc) => {
-      const commentDocRef = doc(db, "cmnt_review", commentDoc.id);
-      await deleteDoc(commentDocRef);
-    });
+      // Use Promise.all() to execute both queries in parallel
+      const [commentSnapshot, replySnapshot] = await Promise.all([
+        getDocs(commentQuery),
+        getDocs(replyQuery),
+      ]);
 
-    // Delete all reply documents
-    replySnapshot.forEach(async (replyDoc) => {
-      const replyDocRef = doc(db, "reply_review", replyDoc.id);
-      await deleteDoc(replyDocRef);
-    });
+      // Delete the post document
+      await deleteDoc(docRef);
 
-    console.log("Post, comments, and replies deleted successfully.");
-    alert("ลบโพสต์สำเร็จ");
-    navigate("/review");
+      // Delete all comment documents
+      commentSnapshot.forEach(async (commentDoc) => {
+        const commentDocRef = doc(db, "cmnt_review", commentDoc.id);
+        await deleteDoc(commentDocRef);
+      });
+
+      // Delete all reply documents
+      replySnapshot.forEach(async (replyDoc) => {
+        const replyDocRef = doc(db, "reply_review", replyDoc.id);
+        await deleteDoc(replyDocRef);
+      });
+
+      // Clear Tag
+      const tagDocRef = doc(db, "tag_ranked", tagName);
+      getDoc(tagDocRef).then((docSnap) => {
+        if (docSnap.exists()) {
+          const tagCount = docSnap.data().count;
+          if (tagCount - 1 === 0) {
+            deleteDoc(tagDocRef);
+          } else {
+            updateDoc(tagDocRef, {
+              count: tagCount - 1,
+            })
+              .then(() => {
+                console.log("Document updated with new count value");
+              })
+              .catch((error) => {
+                console.error("Error updating document: ", error);
+              });
+          }
+        }
+      });
+
+      console.log("Post, comments, and replies deleted successfully.");
+      alert("ลบโพสต์สำเร็จ");
+    }
   };
 
   return (
-    <div>
+    <Dropdown.Menu>
+      <Dropdown.Item onClick={handleReportClick}>รายงานคอมเมนท์</Dropdown.Item>
       {authorCheck ? (
         <div />
       ) : (
-        <Dropdown.Menu>
-          <Dropdown.Item onClick={handleReportClick}>รายงานโพสต์</Dropdown.Item>
-          <Dropdown.Item onClick={handleDeleteClick}>ลบโพสต์</Dropdown.Item>
-        </Dropdown.Menu>
+        <Dropdown.Item onClick={handleDeleteClick}>ลบคอมเมนท์</Dropdown.Item>
       )}
-    </div>
+    </Dropdown.Menu>
   );
 }
 
@@ -1242,80 +1299,90 @@ function Rep_Del_Comment_Click({
       } else {
         setReportedByCurrentUser(false);
       }
-
-      if (member_id === currentUserId) {
-        setAuthorCheck(true);
-      } else {
-        setAuthorCheck(false);
-      }
     } catch (e) {
       console.error(e);
     }
   }, [rep_users, reportedByCurrentUser]);
 
+  useEffect(() => {
+    if (member_id === currentUserId) {
+      setAuthorCheck(false);
+    } else if (member_id !== currentUserId) {
+      setAuthorCheck(true);
+    }
+  }, []);
+
   const handleReportClick = () => {
-    const docRef = doc(db, "cmnt_review", cmnt_id);
-    if (reportedByCurrentUser === false) {
-      updateDoc(docRef, {
-        rep_users: arrayUnion(currentUserId),
-        report: rep_count + 1,
-      })
-        .then(() => {
-          console.log("You Report the post!");
+    const confirmed = window.confirm(
+      "คุณยืนยันที่ต้องการจะรายงานคอมเมนท์ใช่หรือไม่ ?"
+    );
+
+    if (confirmed) {
+      const docRef = doc(db, "cmnt_review", cmnt_id);
+      if (reportedByCurrentUser === false) {
+        updateDoc(docRef, {
+          rep_users: arrayUnion(currentUserId),
+          report: rep_count + 1,
         })
-        .catch((error) => {
-          console.error("Error updating document: ", error);
-        });
-      alert("ขอบคุณที่แจ้งรายงาน ทางแอดมินจะพยายามตรวจสอบให้เร็วที่สุด");
-    } else if (reportedByCurrentUser === true) {
-      alert("คุณได้รายงานคอมเมนท์ไปแล้ว");
+          .then(() => {
+            console.log("You Report the post!");
+          })
+          .catch((error) => {
+            console.error("Error updating document: ", error);
+          });
+        alert("ขอบคุณที่แจ้งรายงาน ทางแอดมินจะพยายามตรวจสอบให้เร็วที่สุด");
+      } else if (reportedByCurrentUser === true) {
+        alert("คุณได้รายงานคอมเมนท์ไปแล้ว");
+      }
     }
   };
 
   const handleDeleteClick = async () => {
-    const postRef = doc(db, "review", postID);
-
-    updateDoc(postRef, {
-      comment: cmnt_count - 1,
-    });
-
-    const docRef = doc(db, "cmnt_review", cmnt_id);
-
-    // Create a query to get all replies for the post
-    const replyQuery = query(
-      collection(db, "reply_review"),
-      where("cmnt_id", "==", cmnt_id)
+    const confirmed = window.confirm(
+      "คุณยืนยันที่ต้องการจะลบคอมเมนท์ใช่หรือไม่ ?"
     );
 
-    // Use Promise.all() to execute both queries in parallel
-    const [replySnapshot] = await Promise.all([getDocs(replyQuery)]);
+    if (confirmed) {
+      const postRef = doc(db, "review", postID);
 
-    // Delete the post document
-    await deleteDoc(docRef);
+      updateDoc(postRef, {
+        comment: cmnt_count - 1,
+      });
 
-    // Delete all reply documents
-    replySnapshot.forEach(async (replyDoc) => {
-      const replyDocRef = doc(db, "reply_review", replyDoc.id);
-      await deleteDoc(replyDocRef);
-    });
+      const docRef = doc(db, "cmnt_review", cmnt_id);
 
-    console.log("Comments, and replies deleted successfully.");
-    alert("ลบคอมเมนท์สำเร็จ");
+      // Create a query to get all replies for the post
+      const replyQuery = query(
+        collection(db, "reply_review"),
+        where("cmnt_id", "==", cmnt_id)
+      );
+
+      // Use Promise.all() to execute both queries in parallel
+      const [replySnapshot] = await Promise.all([getDocs(replyQuery)]);
+
+      // Delete the post document
+      await deleteDoc(docRef);
+
+      // Delete all reply documents
+      replySnapshot.forEach(async (replyDoc) => {
+        const replyDocRef = doc(db, "reply_review", replyDoc.id);
+        await deleteDoc(replyDocRef);
+      });
+
+      console.log("Comments, and replies deleted successfully.");
+      alert("ลบคอมเมนท์สำเร็จ");
+    }
   };
 
   return (
-    <div>
+    <Dropdown.Menu>
+      <Dropdown.Item onClick={handleReportClick}>รายงานคอมเมนท์</Dropdown.Item>
       {authorCheck ? (
         <div />
       ) : (
-        <Dropdown.Menu>
-          <Dropdown.Item onClick={handleReportClick}>
-            รายงานคอมเมนท์
-          </Dropdown.Item>
-          <Dropdown.Item onClick={handleDeleteClick}>ลบคอมเมนท์</Dropdown.Item>
-        </Dropdown.Menu>
+        <Dropdown.Item onClick={handleDeleteClick}>ลบคอมเมนท์</Dropdown.Item>
       )}
-    </div>
+    </Dropdown.Menu>
   );
 }
 
@@ -1340,65 +1407,75 @@ function Rep_Del_Reply_Click({
       } else {
         setReportedByCurrentUser(false);
       }
-
-      if (member_id === currentUserId) {
-        setAuthorCheck(true);
-      } else {
-        setAuthorCheck(false);
-      }
     } catch (e) {
       console.error(e);
     }
   }, [rep_users, reportedByCurrentUser]);
 
+  useEffect(() => {
+    if (member_id === currentUserId) {
+      setAuthorCheck(false);
+    } else if (member_id !== currentUserId) {
+      setAuthorCheck(true);
+    }
+  }, []);
+
   const handleReportClick = () => {
-    const docRef = doc(db, "reply_review", reply_id);
-    if (reportedByCurrentUser === false) {
-      updateDoc(docRef, {
-        rep_users: arrayUnion(currentUserId),
-        report: rep_count + 1,
-      })
-        .then(() => {
-          console.log("You Report the post!");
+    const confirmed = window.confirm(
+      "คุณยืนยันที่ต้องการจะรายงานตอบกลับใช่หรือไม่ ?"
+    );
+
+    if (confirmed) {
+      const docRef = doc(db, "reply_review", reply_id);
+      if (reportedByCurrentUser === false) {
+        updateDoc(docRef, {
+          rep_users: arrayUnion(currentUserId),
+          report: rep_count + 1,
         })
-        .catch((error) => {
-          console.error("Error updating document: ", error);
-        });
-      alert("ขอบคุณที่แจ้งรายงาน ทางแอดมินจะพยายามตรวจสอบให้เร็วที่สุด");
-    } else if (reportedByCurrentUser === true) {
-      alert("คุณได้รายงานตอบกลับไปแล้ว");
+          .then(() => {
+            console.log("You Report the post!");
+          })
+          .catch((error) => {
+            console.error("Error updating document: ", error);
+          });
+        alert("ขอบคุณที่แจ้งรายงาน ทางแอดมินจะพยายามตรวจสอบให้เร็วที่สุด");
+      } else if (reportedByCurrentUser === true) {
+        alert("คุณได้รายงานตอบกลับไปแล้ว");
+      }
     }
   };
 
   const handleDeleteClick = async () => {
-    const cmntRef = doc(db, "cmnt_review", cmnt_id);
+    const confirmed = window.confirm(
+      "คุณยืนยันที่ต้องการจะลบตอบกลับใช่หรือไม่ ?"
+    );
 
-    updateDoc(cmntRef, {
-      reply: reply_count - 1,
-    });
+    if (confirmed) {
+      const cmntRef = doc(db, "cmnt_review", cmnt_id);
 
-    const docRef = doc(db, "reply_review", reply_id);
+      updateDoc(cmntRef, {
+        reply: reply_count - 1,
+      });
 
-    // Delete the post document
-    await deleteDoc(docRef);
+      const docRef = doc(db, "reply_review", reply_id);
 
-    console.log("Replies deleted successfully.");
-    alert("ลบตอบกลับสำเร็จ");
+      // Delete the post document
+      await deleteDoc(docRef);
+
+      console.log("Replies deleted successfully.");
+      alert("ลบตอบกลับสำเร็จ");
+    }
   };
 
   return (
-    <div>
+    <Dropdown.Menu>
+      <Dropdown.Item onClick={handleReportClick}>รายงานคอมเมนท์</Dropdown.Item>
       {authorCheck ? (
         <div />
       ) : (
-        <Dropdown.Menu>
-          <Dropdown.Item onClick={handleReportClick}>
-            รายงานตอบกลับ
-          </Dropdown.Item>
-          <Dropdown.Item onClick={handleDeleteClick}>ลบตอบกลับ</Dropdown.Item>
-        </Dropdown.Menu>
+        <Dropdown.Item onClick={handleDeleteClick}>ลบคอมเมนท์</Dropdown.Item>
       )}
-    </div>
+    </Dropdown.Menu>
   );
 }
